app/requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fuzzywuzzy==0.18.0
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.20.8
lxml==4.9.2
MarkupSafe==2.1.1
psycopg2==2.9.5
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
python-Levenshtein==0.20.8
rapidfuzz==2.13.6
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
urllib3==1.26.13
Werkzeug==2.2.2
